feat: merge mouoa branch

+++ components/ModelTable.js
@@ -9,23 +9,39 @@ import ModelFormDialog from "./ModelFormDialog";
 import ThemedTable from "./ThemedTable";
 import { supplyModelValues } from "./ModelDataView";
 import capitalize from "@/utils/capitalize";
-import { addClassToColumns, supplyValue } from "./Table";
-import { noop } from "@/utils/none";
+import { addClassToColumns, addHeaderClass, supplyValue } from "./Table";
+import { None, noop } from "@/utils/none";
 
+/**
+ *
+ * @param {Object} props
+ * @param {import("@/models/lib/model").Model<Item>} props.Model
+ * @returns
+ */
 export default function ModelTable({
   Model,
-  props = Object.keys(Model.Meta).filter((e) => e[0] !== "!"),
-  headers = props.map((e) => capitalize(Model.Meta[e].label)),
+  Query = Model.all(),
+  props = Model.fields(),
+  headers = props.map((e) => capitalize(Model?.Meta?.[e]?.label ?? e)),
   modelName = sentenceCase(Model.uniqueName()),
   addActionTitle = "Create " + singular(modelName),
   allowEdit = true,
   allowDelete = true,
+  allowCreate = true,
+  enablePrint = false,
+  onClickRow = noop,
   pluralTitle = sentenceCase(modelName),
   onCreate,
+  renderHooks = [],
+  deps = [],
 }) {
-  const { data: items, pager } = useQuery(() => Model.all().pageSize(10), [], {
-    watch: true,
-  });
+  const { data: items, pager } = useQuery(
+    () => Query?.pageSize?.(10),
+    [...deps],
+    {
+      watch: true,
+    }
+  );
   const [formVisible, setFormVisible] = useState(false);
   const [item, setItem] = useState(null);
   useEffect(() => {
@@ -38,7 +54,7 @@ export default function ModelTable({
     setFormVisible(true);
   };
   const createItem = useMutex(async () => {
-    if (allowEdit) {
+    if (allowCreate) {
       const item = (await onCreate?.()) ?? null;
       if (item !== false) {
         setItem(item);
@@ -51,7 +67,7 @@ export default function ModelTable({
   );
   return (
     <>
-      {allowEdit ? (
+      {allowCreate ? (
         <ModelFormDialog
           isOpen={formVisible}
           edit={item}
@@ -61,49 +77,84 @@ export default function ModelTable({
       ) : null}
 
       <Box className="px-4 sm:px-8 py-8">
-        <div className="flex flex-wrap justify-between">
+        <div className="flex flex-wrap justify-between mx-2">
           <Typography variant="h6" as="h2">
             {pluralTitle}
           </Typography>
         </div>
-        <div className="flex flex-wrap pt-6 -mx-2 justify-center">
-          <Button variant="contained" size="large" onClick={createItem}>
-            {addActionTitle} <Add size={32} className="ml-2" />
-          </Button>
-        </div>
+        {allowCreate ? (
+          <div className="flex flex-wrap pt-0 justify-end mx-2">
+            <Button
+              variant="contained"
+              size="large"
+              onClick={createItem}
+              disabled={!Query}
+            >
+              {addActionTitle} <Add size={32} className="ml-2" />
+            </Button>
+          </div>
+        ) : null}
         <ThemedTable
-          title={pluralTitle}
+          // title={pluralTitle}
+
           headers={headers.concat(actions.map(() => ""))}
-          results={items}
+          results={Query ? items : []}
           pager={pager}
-          onClickRow={noop /*(_, row) => showModal(row)*/}
+          enablePrint={enablePrint}
+          onClickRow={(_, row) => onClickRow(items[row])}
+          rowProps={{
+            ...(onClickRow !== noop
+              ? {
+                  sx: {
+                    cursor: "pointer",
+                    "&:hover": {
+                      backgroundColor: "gray.light",
+                    },
+                  },
+                }
+              : None),
+          }}
           renderHooks={[
-            supplyModelValues(props),
+            addHeaderClass("pr-4"),
             addClassToColumns("w-0", [props.length, props.length + 1]),
+            supplyModelValues(props),
+
             supplyValue((row, col, data) => {
+              if (!data) return;
               col -= props.length;
               switch (actions[col]) {
                 case "e":
                   return (
-                    <IconButton color="primary" onClick={() => showModal(row)}>
-                      <Edit />
-                    </IconButton>
+                    <div className="print:hidden">
+                      <IconButton
+                        color="primary"
+                        onClick={() => {
+                          showModal(row);
+                        }}
+                      >
+                        <Edit />
+                      </IconButton>
+                    </div>
                   );
                 case "d":
                   return (
-                    <IconButton
-                      color="error"
-                      onClick={async () => {
-                        if (await confirm("Delete selected item")) {
-                          await data[row].delete();
-                        }
-                      }}
-                    >
-                      <Trash />
-                    </IconButton>
+                    <div className="print:hidden">
+                      <IconButton
+                        color="error"
+                        onClick={async () => {
+                          if (await confirm("Delete selected item")) {
+                            await data[row].delete();
+                          }
+                        }}
+                      >
+                        <Trash />
+                      </IconButton>
+                    </div>
                   );
               }
+              return data;
             }),
+            ...renderHooks,
           ]}
         />
       </Box>
